+++ RELEASES.md
@@ -1,5 +1,21 @@
 # Confidential Match Source Code Releases
 
+## Jul 23, 2026
+
+* Version: v1.508.0
+* Lookup Service
+  * Image Digest: sha256:af0d9667c0862f2a29ec4a5dead7614ee87ccdf2987c91af56e66109a98480e0
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/lookup_server_gcp_signature@sha256:4024c36f497420b0239cbe566b891926a40b1c507a6f76264f1f7351533ecc23
+    )
+* Match Request Processor
+  * Image Digest: sha256:34fa4dd71cb16a767d66321f805d4d0f1c811d5cc78777885b751ff30da53952
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/mrp_app_gcp_signature@sha256:a30922ccaab07aaf720479898f582b6b69d22e7be26f22c132ef21e6b8e5e1de
+    )
+* Match Service
+  * Image Digest: sha256:d276f758f5ddfcce78a851ca28706fac20822a12949806611d54bbfcc9881bff
+  * [Signature](https://us-docker.pkg.dev/admcloud-cfm-public/docker-repo-signatures/match_service_gcp_signature@sha256:5c1d71f787a14956d782496993bc1281bd3ed5f4138236861d235c36c3986c7a
+    )
+
 ## Jul 16, 2026
 
 * Version: v1.502.0